Added MVCComment controller; added view in post.html of the comments; implemented edit, delete and new comment in post.html view file.

+++ src/main/java/com/example/forummanagementsystem/services/CommentServiceImpl.java
@@ -1,8 +1,7 @@
 package com.example.forummanagementsystem.services;
 
 import com.example.forummanagementsystem.exceptions.AuthorizationException;
-import com.example.forummanagementsystem.exceptions.EntityDuplicateException;
-import com.example.forummanagementsystem.exceptions.EntityNotFoundException;
+import com.example.forummanagementsystem.exceptions.BlockedUserException;
 import com.example.forummanagementsystem.models.Comment;
 
 import java.util.List;
@@ -20,6 +19,7 @@ import com.example.forummanagementsystem.repositories.CommentRepository;
 public class CommentServiceImpl implements CommentService {
 
     private static final String MODIFY_COMMENT_ERROR_MESSAGE = "Only admin or comment creator can modify a comment.";
+    public static final String BLOCKED_USER_ERROR_FOR_COMMENTS = "Blocked users are not allowed to create comments.";
 
     private final CommentRepository repository;
 
@@ -40,38 +40,21 @@ public class CommentServiceImpl implements CommentService {
     }
 
 
-
     @Override
     public void create(Comment comment, User user, Post post, Long id) {
-        boolean duplicateExists = false;
-        try {
-            repository.create(comment);
-        } catch (EntityNotFoundException e) {
-            duplicateExists = true;
-        }
-
-        if (duplicateExists) {
-            throw new EntityDuplicateException("Comment", "id", comment.getContent());
-        }
         comment.setUser(user);
         comment.setPost(post);
+        if (comment.getUser().isBlocked()) {
+            throw new BlockedUserException(BLOCKED_USER_ERROR_FOR_COMMENTS);
+        }
+        repository.create(comment);
     }
 
     @Override
     public void update(Comment comment, User user) {
         checkModifyPermissions(comment.getCommentId(), user);
-
-        boolean duplicateExists = true;
-        try {
-            Comment existingComment = repository.getById(comment.getCommentId());
-            if (existingComment.getCommentId() == comment.getCommentId()) {
-                duplicateExists = false;
-            }
-        } catch (EntityNotFoundException e) {
-            duplicateExists = false;
-        }
-        if (duplicateExists) {
-            throw new EntityDuplicateException("Comment", "id", comment.getContent());
+        if (comment.getUser().isBlocked()) {
+            throw new BlockedUserException(BLOCKED_USER_ERROR_FOR_COMMENTS);
         }
         repository.update(comment);
     }
@@ -79,13 +62,14 @@ public class CommentServiceImpl implements CommentService {
     @Override
     public void delete(int id, User user) {
         checkModifyPermissions(id, user);
+        if (getById(id).getUser().isBlocked()) {
+            throw new BlockedUserException(BLOCKED_USER_ERROR_FOR_COMMENTS);
+        }
         repository.delete(id);
     }
 
-
-
     @Override
-    public List<Comment> filter(Optional<String> content, Optional<Integer> commentId,Optional<Integer> postId,
+    public List<Comment> filter(Optional<String> content, Optional<Integer> commentId, Optional<Integer> postId,
                                 Optional<Integer> userId, Optional<String> sort) {
         return repository.filter(content, commentId, postId, userId, sort);
     }
@@ -93,6 +77,10 @@ public class CommentServiceImpl implements CommentService {
     @Override
     public List<Comment> getAll(Optional<String> search) {
         return repository.getAll(search);
+    }
+
+    public List<Comment> getAll() {
+        return repository.getAll();
     }
 
     private void checkModifyPermissions(int commentId, User user) {
